+++ civic-tech-club/src/pages/Home/Hero.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import Container from 'react-bootstrap/Container';
 import Row from 'react-bootstrap/Row';
 import Col from 'react-bootstrap/Col';
-import BlobImage from './BlobImage'
+import BlobImage from '../../components/BlobImage'
 
 const Hero = () => {
     return (
